Opslaan domeintabel

Met deze functie is het mogelijk om een domeintabel op te slaan als .xlsx-bestand of als .csv-bestand.

dom_save(
  naam,
  bestandsnaam = paste(Sys.Date(), naam),
  map = NULL,
  bestandstype = c("xlsx", "csv"),
  peildatum = Sys.Date()
)

Arguments

naam
Naam van een domeintabel - De namen zijn niet hoofdlettergevoelig. Zie dom_overzicht() voor geldige domeintabelnamen.

bestandsnaam
Naam van het bestand om op te slaan. Eventueel als volledig pad. Default is de datum gevolgd door de naam van de domeintabel. De bestandsnaam mag zonder extensie worden opgegeven.

map
Naam van de map. De map moet bestaan. Optioneel.

bestandstype
"xlsx" of "csv" Opslaan als xlsx of als csv-bestand. Default is xlsx.

peildatum
Date of een character die omgezet kan worden in een Date met lubridate::as_date(). De peildatum filtert de output om alleen geldige domeinwaarden op de peildatum weer te geven. Gebruik NULL om alle domeinwaarden ongeacht de geldigheid weer te geven.

Value

Slaat de domeintabel op de schijf. De domeintabel zelf wordt onzichtbaar geretourneerd.

Details

Voor het opslaan van een bestand als .xlsx wordt openxlsx::write.xlsx() gebruikt. Voor het opslaan als .csv wordt readr::write_csv2() gebruikt.

Opgeslagen excelbestanden kunnen ingelezen worden met openxlsx::read.xlsx() of readxl::read_excel(). csv-bestanden kunnen worden ingelezen met readr::read_csv2()

See also

Examples

if (FALSE) {
 dom_save("MonsterType")
 dom_save("MonsterType", bestandsnaam = "test.csv", map = "data", bestandstype = "csv")

}